requirements.txt
discord.py
PyYAML
python-dotenv
yt-dlp
Pillow
aiohttp
